cli/commands/build-cache: simplify build and move cmdOutput to helpers as runPlain
